Main-To-App Cleanup
