+++ confidential-app/src/util/urls.ts
@@ -1,0 +1,6 @@
+export enum ROUTE_PATH {
+    HOME = '/',
+    LOGIN = '/login',
+    REGISTER = '/register',
+    USER_PROFILE = '/user-profile'
+}
